+++ tests/Fixtures/TestRecord.php
@@ -29,8 +29,6 @@ class TestRecord
 
   public static function resetFixture(): void
   {
-    static::$_schema = null;
-    static::$_schema_connection = null;
     static::$_static_conn = null;
   }